+++ uefi/src/proto/pci/mod.rs
@@ -109,8 +109,15 @@ impl PartialOrd for PciIoAddress {
 }
 
 impl Ord for PciIoAddress {
-    fn cmp(&self, other: &Self) -> Ordering {
-        u64::from(*self).cmp(&u64::from(*other))
+    fn cmp(&self, o: &Self) -> Ordering {
+        // extract fields because taking references to unaligned fields in packed structs is a nono
+        let (bus, dev, fun, reg, ext_reg) = (self.bus, self.dev, self.fun, self.reg, self.ext_reg);
+        let (o_bus, o_dev, o_fun, o_reg, o_ext_reg) = (o.bus, o.dev, o.fun, o.reg, o.ext_reg);
+        bus.cmp(&o_bus)
+            .then(dev.cmp(&o_dev))
+            .then(fun.cmp(&o_fun))
+            .then(reg.cmp(&o_reg))
+            .then(ext_reg.cmp(&o_ext_reg))
     }
 }
 
@@ -152,6 +159,8 @@ fn encode_io_mode_and_unit<U: PciIoUnit>(mode: PciIoMode) -> PciRootBridgeIoProt
 
 #[cfg(test)]
 mod tests {
+    use core::cmp::Ordering;
+
     use super::PciIoAddress;
 
     #[test]
@@ -169,5 +178,44 @@ mod tests {
         let dstaddr = PciIoAddress::from(rawaddr);
         assert_eq!(rawaddr, 0x99_bb_dd_ff_7755_3311);
         assert_eq!(srcaddr, dstaddr);
+    }
+
+    #[test]
+    fn test_pci_order() {
+        let addr0_0_0 = PciIoAddress::new(0, 0, 0);
+        let addr0_0_1 = PciIoAddress::new(0, 0, 1);
+        let addr0_1_0 = PciIoAddress::new(0, 1, 0);
+        let addr1_0_0 = PciIoAddress::new(1, 0, 0);
+
+        assert_eq!(addr0_0_0.cmp(&addr0_0_0), Ordering::Equal);
+        assert_eq!(addr0_0_0.cmp(addr0_0_1), Ordering::Less);
+        assert_eq!(addr0_0_0.cmp(&addr0_1_0), Ordering::Less);
+        assert_eq!(addr0_0_0.cmp(&addr1_0_0), Ordering::Less);
+
+        assert_eq!(addr0_0_1.cmp(addr0_0_0), Ordering::Greater);
+        assert_eq!(addr0_0_1.cmp(addr0_0_1), Ordering::Equal);
+        assert_eq!(addr0_0_1.cmp(&addr0_1_0), Ordering::Less);
+        assert_eq!(addr0_0_1.cmp(&addr1_0_0), Ordering::Less);
+
+        assert_eq!(addr0_1_0.cmp(addr0_0_0), Ordering::Greater);
+        assert_eq!(addr0_1_0.cmp(addr0_0_1), Ordering::Greater);
+        assert_eq!(addr0_1_0.cmp(&addr0_1_0), Ordering::Equal);
+        assert_eq!(addr0_1_0.cmp(&addr1_0_0), Ordering::Less);
+
+        assert_eq!(addr1_0_0.cmp(addr0_0_0), Ordering::Greater);
+        assert_eq!(addr1_0_0.cmp(addr0_0_1), Ordering::Greater);
+        assert_eq!(addr1_0_0.cmp(&addr0_1_0), Ordering::Greater);
+        assert_eq!(addr1_0_0.cmp(&addr1_0_0), Ordering::Equal);
+
+        assert_eq!(addr0_0_0.cmp(addr0_0_0.with_register(1)), Ordering::Less);
+        assert_eq!(addr0_0_0.with_register(1).cmp(addr0_0_0), Ordering::Greater);
+        assert_eq!(
+            addr0_0_0.cmp(addr0_0_0.with_extended_register(1)),
+            Ordering::Less
+        );
+        assert_eq!(
+            addr0_0_0.with_extended_register(1).cmp(addr0_0_0),
+            Ordering::Greater
+        );
     }
 }