+++ uefi/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added `proto::shell::Shell::{var(), set_var(), vars()}`
 
 ## Changed
-
+- Changed ordering of `proto::pci::PciIoAddress` to (bus -> dev -> fun -> reg -> ext_reg).
 
 # uefi - v0.36.1 (2025-11-05)